+++ packages/color-studio/builders/sketchpalette-from-json.js
@@ -1,16 +1,13 @@
-const flatten = require('lodash/flatten')
-
 const PALETTE = require('../dist/colors.json')
 
-const print = require('../utilities/print-json')
 const toSketchPalette = require('../utilities/to-sketch-palette')
 
 const colors = PALETTE.colors.map(colorArray => {
   return colorArray.map(colorObject => colorObject.value)
 })
 
-const palette = toSketchPalette(flatten(colors), {
+const palette = toSketchPalette(colors, {
   _paletteVersion: PALETTE.version
 })
 
-print(palette)
+console.log(palette)
