+++ packages/color-studio/data/base-colors.js
@@ -1,4 +1,4 @@
-const chroma = require('chroma-js')
+const { colden, hotten, mix } = require('../utilities/base-color-helpers')
 
 const values = {
   gray: '#707070',
@@ -98,15 +98,3 @@ module.exports = [
     formula: 'secondary'
   }
 ]
-
-function mix(color1, color2, ratio = 0.5) {
-  return chroma.mix(color1, color2, ratio, 'lch').hex()
-}
-
-function hotten(color) {
-  return chroma(color).darken(0.2).saturate(2).hex()
-}
-
-function colden(color) {
-  return chroma(color).set('hsl.s', '*0.75').desaturate(0.2).hex()
-}
