+++ packages/color-studio/docs/custom.html
@@ -22,6 +22,6 @@
 
     <div id="color-tiles"></div>
 
-    <script src="build/main.js"></script>
+    <script src="build/custom.js"></script>
   </body>
 </html>
