+++ packages/color-studio/docs/index.html
@@ -27,6 +27,6 @@
 
     <div id="color-tiles"></div>
 
-    <script src="build/main.js"></script>
+    <script src="build/index.js"></script>
   </body>
 </html>
