+++ packages/color-studio/package.json
@@ -17,8 +17,10 @@
     "mini-css-extract-plugin": "^0.4.0",
     "node-sass": "^4.8.3",
     "postcss-loader": "^2.1.3",
+    "raw-loader": "^0.5.1",
     "sass-loader": "^6.0.7",
-    "webpack": "4.19.1"
+    "webpack": "4.19.1",
+    "webpack-extraneous-file-cleanup-plugin": "^2.0.0"
   },
   "devDependencies": {
     "husky": "^0.14.3",
@@ -35,7 +37,7 @@
     "precommit": "yarn lint && yarn build && git add --update",
     "studio": "yarn studio:build",
     "studio:build": "webpack --display=errors-only",
-    "studio:development": "webpack --watch --mode=development"
+    "studio:development": "webpack --watch"
   },
   "xo": {
     "ignores": [
