--- packages/color-studio/utilities/print-json.js
@@ -1,3 +1,0 @@
-module.exports = data => {
-  console.log(JSON.stringify(data, null, 2))
-}
